Refactor: ErrorResponse 구조 변경
+ ValidationError에 rejectedValue 제공
